Add relationship between applicant and resume model

+++ src/main/java/com/inverse/project/Jobless/services/impl/ResumeServiceImpl.java
@@ -3,7 +3,9 @@ package com.inverse.project.Jobless.services.impl;
 import com.inverse.project.Jobless.config.ModelMap;
 import com.inverse.project.Jobless.dto.ResumeDto;
 import com.inverse.project.Jobless.exceptions.ResourceNotFoundException;
+import com.inverse.project.Jobless.models.Applicant;
 import com.inverse.project.Jobless.models.Resume;
+import com.inverse.project.Jobless.repositories.ApplicantRepository;
 import com.inverse.project.Jobless.repositories.ResumeRepository;
 import com.inverse.project.Jobless.services.ResumeService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,10 +16,17 @@ public class ResumeServiceImpl implements ResumeService {
     @Autowired
     private ResumeRepository repository;
     @Autowired
+    private ApplicantRepository applicantRepository;
+    @Autowired
     private ModelMap modelMap;
     @Override
-    public ResumeDto create(ResumeDto resumeDto) {
+    public ResumeDto create(ResumeDto resumeDto, Integer applicantId) {
         Resume resume = this.modelMap.modelMapper().map(resumeDto, Resume.class);
+        Applicant applicant = this.applicantRepository.findById(applicantId)
+                .orElseThrow(
+                        () -> new ResourceNotFoundException("Applicant not found: ID " + applicantId)
+                );
+        resume.setApplicant(applicant);
         this.repository.save(resume);
         return this.modelMap.modelMapper().map(resume, ResumeDto.class);
     }
@@ -31,11 +40,14 @@ public class ResumeServiceImpl implements ResumeService {
     }
 
     @Override
-    public ResumeDto update(ResumeDto resumeDto, Integer id) {
+    public ResumeDto update(ResumeDto resumeDto,Integer applicantId, Integer id) {
         Resume resume = this.repository.findById(id)
                 .orElseThrow(
                         () -> new ResourceNotFoundException("Resume not found: ID " + id)
                 );
+        if (!resume.getApplicant().getId().equals(applicantId)) {
+            throw new ResourceNotFoundException("Applicant Id not found: " + applicantId);
+        }
         resume.setName(resumeDto.getName());
         resume.setEmail(resumeDto.getEmail());
         resume.setPhone(resumeDto.getPhone());
@@ -48,11 +60,14 @@ public class ResumeServiceImpl implements ResumeService {
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Integer applicantId,Integer id) {
         Resume resume = this.repository.findById(id)
                 .orElseThrow(
                         () -> new ResourceNotFoundException("Resume not found: ID " + id)
                 );
+        if (!resume.getApplicant().getId().equals(applicantId)) {
+            throw new ResourceNotFoundException("Applicant Id not found: " + applicantId);
+        }
         this.repository.delete(resume);
     }
 }
